+++ test/10-http.spec.js
@@ -4,13 +4,13 @@
 
 const { assert } = require('chai');
 const nock = require('nock');
+const axios = require('./../src/axios');
 const HttpClient = require('./../src/http');
 const pkg = require('./../package.json');
 
 
 describe('http', () => {
     let testClient;
-    let testClient2;
 
     const defaultUserAgent = `node-${pkg.name}/${pkg.version}`;
     const customUserAgent = 'custom-ua-123';
@@ -49,6 +49,10 @@ describe('http', () => {
             .reply(200, 'ok');
     });
 
+    after(() => {
+        axios.defaults.headers.common['User-Agent'] = defaultUserAgent;
+    });
+
 
     /**
      * Initialize
@@ -56,12 +60,6 @@ describe('http', () => {
 
     it('should initialize clients', () => {
         testClient = new HttpClient();
-
-        testClient2 = new HttpClient(null, null, {
-            headers: {
-                'User-Agent': customUserAgent
-            }
-        });
     });
 
 
@@ -83,6 +81,7 @@ describe('http', () => {
 
     it('should request using default user-agent', async () => {
         const resp = await testClient.request('http://default-ua.example.com', 'get');
+
         assert.isObject(resp);
         assert.strictEqual(resp.status, 200);
         assert.strictEqual(resp.data, 'ok');
@@ -93,13 +92,16 @@ describe('http', () => {
     });
 
     it('should request using custom user-agent', async () => {
-        const resp = await testClient2.request('http://custom-ua.example.com', 'get');
+        axios.defaults.headers.common['User-Agent'] = customUserAgent;
+        const resp = await testClient.request('http://custom-ua.example.com', 'get');
+
         assert.isObject(resp);
         assert.strictEqual(resp.status, 200);
         assert.strictEqual(resp.data, 'ok');
     });
 
     it('should not request using default user-agent', async () => {
-        await assert.isRejected(testClient2.request('http://default-ua.example.com', 'get'));
+        axios.defaults.headers.common['User-Agent'] = customUserAgent;
+        await assert.isRejected(testClient.request('http://default-ua.example.com', 'get'));
     });
 });