+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 ## v3.0.0
 
-* `added` Allow passing `axios` config through to the HTTP client
+* `added` Expose `axios` instance to allow manipulating HTTP client defaults
 * `breaking` Remove support for Node v4 and Node v6
 * `breaking` Remove Babel transpilation